+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Vz Keymap extension will be documented in this file.
 
-### [Unreleased]
+### [0.10.2] - 2020-08-02
 - Improved:
   - The behavior of Paste of a text which is from box-selection mode.
   - The behavior of Paste of a text which has been taken from the last line of a document.